requirements-cpu.txt
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.3
plotly==5.3.1
pystan==2.19.1.1
prophet==1.0.1
scikit_learn==1.0.1
tensorflow-cpu
yfinance==0.1.63
pytest
coverage